Switch lib and main to make doc check work
